% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec_anomaly.R
\name{prec_anomaly}
\alias{prec_anomaly}
\title{Precipitation anomaly in the last 100 days for a given weather station}
\usage{
prec_anomaly(station, days = 100)
}
\arguments{
\item{station}{Station name from DWD (German Weather Service)}

\item{days}{How many days (backweards from today) should be considered? Delaut = 100 days.}
}
\value{
A table. tbc.
}
\description{
Precipitation anomaly in the last 100 days for a given weather station
}
